Provide a method for derived contracts to set their own ENS reverse name.
`setReverseName` can be called to set the ENS reverse name in case a derived
contract needs to retain ownership of its reverse root node. It is up to the
derived contract to either call `setReverseName` or release ownership of the
root node with `releaseRootNodeOwnership` for reverse resolution to work as
expected.